Memoized value in AuthContext
Move validate to utils
